+++ server/api/meta.get.ts
@@ -1,0 +1,6 @@
+import { defineEventHandler } from 'h3'
+import { readMetaFromDisk } from '../metaDisk'
+
+export default defineEventHandler(async () => {
+  return readMetaFromDisk()
+})